requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2.1
django-pure-pagination==0.3.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
Pygments==2.4.0
pytz==2019.1
raven==6.10.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
